requirements.txt
discord.py==2.5.2
python-dotenv
psutil==7.0.0